db2-postgres-pipeline/requirements.txt
# Core dependencies
pika>=1.3.0
psycopg2-binary>=2.9.0
pyodbc>=5.0.0
redis>=5.0.0
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.0.0
colorlog>=6.7.0

# Utilities
click>=8.1.0